sources/awesome-privacy-skill/requirements.txt
# File format support
pypdf>=4.0.0
openpyxl>=3.1.0
python-docx>=1.1.0

# Image processing
Pillow>=10.0.0
pytesseract>=0.3.10

# Optional OCR engines (choose one)
# paddleocr>=2.7.0
# rapidocr-onnxruntime>=1.3.0
